feat udistrital/arka_cliente#584: Ajustes
- Archivos individuales por tipo de movimiento al que se realiza el ajuste
