+++ helpers/ajustesHelper/ajustes.helper.go
@@ -3,31 +3,21 @@ package ajustesHelper
 import (
 	"encoding/json"
 	"fmt"
-	"math"
 	"net/url"
-	"strconv"
 	"time"
 
 	"github.com/astaxie/beego"
 	"github.com/astaxie/beego/logs"
-	"github.com/udistrital/arka_mid/helpers/actaRecibido"
-	"github.com/udistrital/arka_mid/helpers/asientoContable"
-	"github.com/udistrital/arka_mid/helpers/catalogoElementosHelper"
 	"github.com/udistrital/arka_mid/helpers/cuentasContablesHelper"
-	"github.com/udistrital/arka_mid/helpers/depreciacionHelper"
-	"github.com/udistrital/arka_mid/helpers/entradaHelper"
 	"github.com/udistrital/arka_mid/helpers/movimientosArkaHelper"
 	"github.com/udistrital/arka_mid/helpers/movimientosContablesMidHelper"
 	"github.com/udistrital/arka_mid/helpers/parametrosHelper"
-	"github.com/udistrital/arka_mid/helpers/salidaHelper"
 	"github.com/udistrital/arka_mid/helpers/tercerosHelper"
 	"github.com/udistrital/arka_mid/helpers/utilsHelper"
 	"github.com/udistrital/arka_mid/models"
 	"github.com/udistrital/utils_oas/errorctrl"
 	"github.com/udistrital/utils_oas/formatdata"
 )
-
-const queryUD string = "query=TipoDocumentoId__Nombre:NIT,Numero:"
 
 func PostAjuste(trContable *models.PreTrAjuste) (movimiento *models.Movimiento, outputError map[string]interface{}) {
 
@@ -281,734 +271,4 @@ func AprobarAjuste(id int) (movimiento *models.Movimiento, outputError map[strin
 	}
 
 	return movimiento, nil
-}
-
-func GenerarAjuste(elementos []*models.DetalleElemento_) (resultado []*models.MovimientoTransaccion, outputError map[string]interface{}) {
-
-	funcion := "GenerarAjuste"
-	defer errorctrl.ErrorControlFunction(funcion+" - Unhandled Error!", "500")
-
-	var (
-		query                string
-		idsEl                []int
-		entrada              *models.Movimiento
-		orgActa              []*models.Elemento
-		elementosSalida      map[int]*models.ElementosPorActualizarSalida
-		// updateMsc            []*models.DetalleElemento_
-		updateVls            []*models.DetalleElemento_
-		updateSg             []*models.DetalleElemento_
-		updateMp             []*models.DetalleElemento_
-		movimientos          []*models.MovimientoTransaccion
-		tipoMovimientoSalida int
-		novedadesMedicion    map[int][]*models.NovedadElemento
-	)
-
-	for _, el := range elementos {
-		idsEl = append(idsEl, el.Id)
-	}
-
-	query = "Id__in:" + utilsHelper.ArrayToString(idsEl, "|")
-	if elementos_, err := actaRecibido.GetAllElemento(query, "", "Id", "desc", "0", "-1"); err != nil {
-		return nil, err
-	} else {
-		orgActa = elementos_
-	}
-
-	if entrada_, err := movimientosArkaHelper.GetEntradaByActa(orgActa[0].ActaRecibidoId.Id); err != nil {
-		return nil, err
-	} else if entrada_ == nil {
-		return nil, nil
-	} else {
-		entrada = entrada_
-	}
-
-	if _, vls, sg, mp, err := separarElementosPorModificacion(orgActa, elementos, entrada.EstadoMovimientoId.Nombre == "Entrada Con Salida"); err != nil {
-		return nil, err
-	} else {
-		// updateMsc = msc
-		updateVls = vls
-		updateSg = sg
-		updateMp = mp
-	}
-
-	if (len(updateVls)+len(updateSg) > 0) && (entrada.EstadoMovimientoId.Nombre == "Entrada Aprobada" || entrada.EstadoMovimientoId.Nombre == "Entrada Con Salida") {
-		var proveedorId int
-		var consecutivo string
-
-		query = "Activo:true,ActaRecibidoId__Id:" + strconv.Itoa(orgActa[0].ActaRecibidoId.Id)
-		if ha, err := actaRecibido.GetAllHistoricoActa(query, "", "FechaCreacion", "desc", "", "-1"); err != nil {
-			return nil, err
-		} else {
-			proveedorId = ha[0].ProveedorId
-		}
-
-		if cs, err := entradaHelper.GetConsecutivoEntrada(entrada.Detalle); err != nil {
-			return nil, err
-		} else {
-			consecutivo = cs
-		}
-
-		if movsEntrada, err := calcularAjusteMovimiento(orgActa, updateVls, updateSg, entrada.FormatoTipoMovimientoId.Id, consecutivo, proveedorId, "Entrada"); err != nil {
-			return nil, err
-		} else {
-			movimientos = append(movimientos, movsEntrada...)
-		}
-	}
-
-	if entrada.EstadoMovimientoId.Nombre == "Entrada Con Salida" {
-
-		query = "limit=-1&sortby=MovimientoId,ElementoActaId&order=desc,desc&query=ElementoActaId__in:" + utilsHelper.ArrayToString(idsEl, "|")
-		if elementos_, err := movimientosArkaHelper.GetAllElementosMovimiento(query); err != nil {
-			return nil, err
-		} else {
-			if elementosSalida_, updateMp_, err := separarElementosPorSalida(elementos_, updateVls, updateSg, updateMp); err != nil {
-				return nil, err
-			} else {
-				elementosSalida = elementosSalida_
-				updateMp = updateMp_
-			}
-
-			if len(elementosSalida) > 0 {
-				query = "query=CodigoAbreviacion:SAL"
-				if fm, err := movimientosArkaHelper.GetAllFormatoTipoMovimiento(query); err != nil {
-					return nil, err
-				} else {
-					tipoMovimientoSalida = fm[0].Id
-				}
-			}
-			idsEl = []int{}
-		}
-	}
-
-	for _, elms := range elementosSalida {
-
-		var funcionario int
-		var consecutivo string
-
-		if func_, cons_, err := salidaHelper.GetInfoSalida(elms.Salida.Detalle); err != nil {
-			return nil, err
-		} else {
-			funcionario = func_
-			consecutivo = cons_
-		}
-
-		for _, el := range elms.UpdateSg {
-			idsEl = append(idsEl, el.Id)
-		}
-
-		for _, el := range elms.UpdateVls {
-			idsEl = append(idsEl, el.Id)
-		}
-
-		if movsSalida, err := calcularAjusteMovimiento(orgActa, elms.UpdateVls, elms.UpdateSg, tipoMovimientoSalida, consecutivo, funcionario, "Salida"); err != nil {
-			return nil, err
-		} else {
-			movimientos = append(movimientos, movsSalida...)
-		}
-
-	}
-
-	if len(idsEl) > 0 {
-		query = "limit=-1&sortby=MovimientoId,FechaCreacion&order=asc,asc&query=ElementoMovimientoId__ElementoActaId__in:" + utilsHelper.ArrayToString(idsEl, "|")
-		if novedades_, err := movimientosArkaHelper.GetAllNovedadElemento(query); err != nil {
-			return nil, err
-		} else {
-			novedadesMedicion = separarNovedadesPorElemento(novedades_)
-
-			if movimientos_, _, err := calcularAjusteMediciones(novedadesMedicion, updateSg, updateVls, updateMp, orgActa); err != nil {
-				return nil, err
-			} else {
-				movimientos = append(movimientos, movimientos_...)
-			}
-		}
-	}
-
-	return movimientos, nil
-}
-
-// separarElementosPorModificacion Separa los elementos según se deba modificar Subgrupo, Valores, Misceláneos o Mediciones posteriores
-func separarElementosPorModificacion(originales []*models.Elemento, actualizados []*models.DetalleElemento_, mediciones bool) (msc, vls, sg, mp []*models.DetalleElemento_, outputError map[string]interface{}) {
-
-	funcion := "separarElementosPorModificacion"
-	defer errorctrl.ErrorControlFunction(funcion+" - Unhandled Error!", "500")
-
-	msc = make([]*models.DetalleElemento_, 0)
-	vls = make([]*models.DetalleElemento_, 0)
-	sg = make([]*models.DetalleElemento_, 0)
-	mp = make([]*models.DetalleElemento_, 0)
-
-	for _, el_ := range originales {
-		if idx := findElementoInArrayD(actualizados, el_.Id); idx > -1 {
-			if msc_, vls_, sg_, err := determinarDeltaActa(el_, actualizados[idx]); err != nil {
-				return nil, nil, nil, nil, err
-			} else if msc_ != nil {
-				msc = append(msc, msc_)
-			} else if vls_ != nil {
-				vls = append(vls, vls_)
-			} else if sg_ != nil {
-				sg = append(sg, sg_)
-			} else if mediciones {
-				mp = append(mp, actualizados[idx])
-			}
-		}
-	}
-
-	return msc, vls, sg, mp, nil
-
-}
-
-// separarElementosPorSalida Separa los elementos según el tipo de ajuste de cada uno y los agrupa según la salida
-func separarElementosPorSalida(elementos []*models.ElementosMovimiento, updateVls, updateSg, updateMp []*models.DetalleElemento_) (elementosSalidas map[int]*models.ElementosPorActualizarSalida, pendientes_ []*models.DetalleElemento_, outputError map[string]interface{}) {
-
-	funcion := "separarElementosPorModificacion"
-	defer errorctrl.ErrorControlFunction(funcion+" - Unhandled Error!", "500")
-
-	elementosSalidas = make(map[int]*models.ElementosPorActualizarSalida)
-	for _, el := range elementos {
-
-		if len(updateMp) > 0 {
-			if idx := findElementoInArrayD(updateMp, el.ElementoActaId); idx > -1 {
-				if updateMp[idx].ValorResidual == el.ValorResidual && updateMp[idx].VidaUtil == el.VidaUtil {
-					updateMp = append(updateMp[:idx], updateMp[idx+1:]...)
-				}
-				continue
-			}
-		} else if el.MovimientoId.EstadoMovimientoId.Nombre != "Salida Aprobada" {
-			continue
-		}
-
-		if len(updateSg) > 0 {
-			if idx := findElementoInArrayD(updateSg, el.ElementoActaId); idx > -1 {
-				if elementosSalidas[el.MovimientoId.Id] == nil {
-					elementosSalidas[el.MovimientoId.Id] = new(models.ElementosPorActualizarSalida)
-					elementosSalidas[el.MovimientoId.Id].Salida = el.MovimientoId
-				}
-
-				elementosSalidas[el.MovimientoId.Id].UpdateSg = append(elementosSalidas[el.MovimientoId.Id].UpdateSg, updateSg[idx])
-				updateSg = append(updateSg[:idx], updateSg[idx+1:]...)
-			}
-		} else if len(updateVls) > 0 {
-			if idx := findElementoInArrayD(updateVls, el.ElementoActaId); idx > -1 {
-				if elementosSalidas[el.MovimientoId.Id] == nil {
-					elementosSalidas[el.MovimientoId.Id] = new(models.ElementosPorActualizarSalida)
-					elementosSalidas[el.MovimientoId.Id].Salida = el.MovimientoId
-				}
-
-				elementosSalidas[el.MovimientoId.Id].UpdateVls = append(elementosSalidas[el.MovimientoId.Id].UpdateVls, updateVls[idx])
-				updateVls = append(updateVls[:idx], updateVls[idx+1:]...)
-			}
-		}
-
-	}
-	return elementosSalidas, updateMp, nil
-
-}
-
-// separarNovedadesPorElemento Separa las novedades por elementos
-func separarNovedadesPorElemento(novedades []*models.NovedadElemento) (novedades_ map[int][]*models.NovedadElemento) {
-
-	novedades_ = make(map[int][]*models.NovedadElemento, 0)
-	for _, nv := range novedades {
-		novedades_[nv.ElementoMovimientoId.Id] = append(novedades_[nv.ElementoMovimientoId.Id], nv)
-	}
-
-	return novedades_
-
-}
-
-// determinarDeltaActa Separa elementos según el ajuste
-func determinarDeltaActa(org *models.Elemento, nvo *models.DetalleElemento_) (msc, vls, sg *models.DetalleElemento_, outputError map[string]interface{}) {
-
-	funcion := "determinarDeltaActa"
-	defer errorctrl.ErrorControlFunction(funcion+" - Unhandled Error!", "500")
-
-	if org.SubgrupoCatalogoId != nvo.SubgrupoCatalogoId {
-
-		urlcrud := "fields=TipoBienId&sortby=Id&order=desc&query=Activo:true,SubgrupoId__Id:" + strconv.Itoa(nvo.SubgrupoCatalogoId)
-		if detalleSubgrupo_, err := catalogoElementosHelper.GetAllDetalleSubgrupo(urlcrud); err != nil {
-			return nil, nil, nil, err
-		} else if len(detalleSubgrupo_) == 0 {
-			err := "len(detalleSubgrupo_) = 0"
-			eval := " - catalogoElementosHelper.GetAllDetalleSubgrupo(urlcrud)"
-			return nil, nil, nil, errorctrl.Error(funcion+eval, err, "500")
-		} else {
-			if detalleSubgrupo_[0].TipoBienId.NecesitaPlaca && nvo.Placa != "" {
-				ctxPlaca, _ := beego.AppConfig.Int("contxtPlaca")
-				if placa_, _, err := utilsHelper.GetConsecutivo("%05.0f", ctxPlaca, "Registro Placa Arka"); err != nil {
-					return nil, nil, nil, err
-				} else {
-					year, month, day := time.Now().Date()
-					nvo.Placa = utilsHelper.FormatConsecutivo(fmt.Sprintf("%04d%02d%02d", year, month, day), placa_, "")
-				}
-			} else if !detalleSubgrupo_[0].TipoBienId.NecesitaPlaca && nvo.Placa != "" {
-				nvo.Placa = ""
-			}
-
-		}
-
-		nvo.Activo = true
-		sg = nvo
-
-	} else if org.ValorTotal != nvo.ValorTotal {
-		nvo.Activo = true
-		vls = nvo
-
-	} else if org.Nombre != nvo.Nombre || org.Marca != nvo.Marca ||
-		org.Serie != nvo.Serie || org.UnidadMedida != nvo.UnidadMedida ||
-		org.Cantidad != nvo.Cantidad || org.ValorUnitario != nvo.ValorUnitario ||
-		org.Subtotal != nvo.Subtotal || org.Descuento != nvo.Descuento ||
-		org.PorcentajeIvaId != nvo.PorcentajeIvaId ||
-		org.ValorIva != nvo.ValorIva || org.ValorFinal != nvo.ValorFinal {
-
-		nvo.Activo = true
-		msc = nvo
-
-	}
-
-	return msc, vls, sg, nil
-
-}
-
-// calcularAjusteMovimiento Calcula la transacción contable generada a partir de los elementos y el cambio de cada uno
-func calcularAjusteMovimiento(originales []*models.Elemento, actualizarVl, actualizarSg []*models.DetalleElemento_, movimientoId int, consecutivo string, proveedorId int, tipoMovimiento string) (movimientos []*models.MovimientoTransaccion, outputError map[string]interface{}) {
-
-	funcion := "calcularAjusteMovimiento"
-	defer errorctrl.ErrorControlFunction(funcion+" - Unhandled Error!", "500")
-
-	var (
-		ids             []int
-		movDebito       int
-		movCredito      int
-		cuentasSubgrupo map[int]*models.CuentaSubgrupo
-		detalleCuenta   map[string]*models.CuentaContable
-	)
-
-	cuentasSubgrupo = make(map[int]*models.CuentaSubgrupo)
-	detalleCuenta = make(map[string]*models.CuentaContable)
-	if db_, cr_, err := parametrosHelper.GetParametrosDebitoCredito(); err != nil {
-		return nil, err
-	} else {
-		movDebito = db_
-		movCredito = cr_
-	}
-
-	for _, el := range originales {
-		ids = append(ids, el.SubgrupoCatalogoId)
-	}
-
-	for _, el := range actualizarSg {
-		ids = append(ids, el.SubgrupoCatalogoId)
-	}
-
-	query := "limit=-1&fields=CuentaDebitoId,CuentaCreditoId,SubgrupoId&sortby=Id&order=desc&"
-	query += "query=SubtipoMovimientoId:" + strconv.Itoa(movimientoId) + ",Activo:true,SubgrupoId__Id__in:"
-	query += url.QueryEscape(utilsHelper.ArrayToString(ids, "|"))
-	if cuentas_, err := catalogoElementosHelper.GetAllCuentasSubgrupo(query); err != nil {
-		return nil, err
-	} else {
-		for _, cuenta := range cuentas_ {
-			cuentasSubgrupo[cuenta.SubgrupoId.Id] = cuenta
-		}
-	}
-
-	dsc := getDescripcionMovContable(tipoMovimiento, consecutivo)
-	for _, el := range actualizarSg {
-		if idx := findElementoInArrayE(originales, el.Id); idx > -1 {
-
-			if detalleCuenta_, err := fillCuentas(detalleCuenta, []string{cuentasSubgrupo[originales[idx].SubgrupoCatalogoId].CuentaCreditoId,
-				cuentasSubgrupo[el.SubgrupoCatalogoId].CuentaCreditoId, cuentasSubgrupo[originales[idx].SubgrupoCatalogoId].CuentaDebitoId, cuentasSubgrupo[el.SubgrupoCatalogoId].CuentaDebitoId}); err != nil {
-				return nil, err
-			} else {
-				detalleCuenta = detalleCuenta_
-			}
-
-			if cuentasSubgrupo[el.SubgrupoCatalogoId].CuentaCreditoId != cuentasSubgrupo[originales[idx].SubgrupoCatalogoId].CuentaCreditoId {
-
-				movimientoR := asientoContable.CreaMovimiento(originales[idx].ValorTotal, dsc, proveedorId, detalleCuenta[cuentasSubgrupo[originales[idx].SubgrupoCatalogoId].CuentaCreditoId], movDebito)
-				movimiento := asientoContable.CreaMovimiento(el.ValorTotal, dsc, proveedorId, detalleCuenta[cuentasSubgrupo[el.SubgrupoCatalogoId].CuentaCreditoId], movCredito)
-				movimientos = append(movimientos, movimientoR, movimiento)
-
-			} else if el.ValorTotal != originales[idx].ValorTotal {
-
-				tipoMovimiento := movCredito
-				if el.ValorTotal < originales[idx].ValorTotal {
-					tipoMovimiento = movDebito
-				}
-
-				movimiento := asientoContable.CreaMovimiento(math.Abs(el.ValorTotal-originales[idx].ValorTotal), dsc, proveedorId, detalleCuenta[cuentasSubgrupo[el.SubgrupoCatalogoId].CuentaCreditoId], tipoMovimiento)
-				movimientos = append(movimientos, movimiento)
-
-			}
-
-			if cuentasSubgrupo[el.SubgrupoCatalogoId].CuentaDebitoId != cuentasSubgrupo[originales[idx].SubgrupoCatalogoId].CuentaDebitoId {
-
-				movimientoR := asientoContable.CreaMovimiento(originales[idx].ValorTotal, dsc, proveedorId, detalleCuenta[cuentasSubgrupo[originales[idx].SubgrupoCatalogoId].CuentaDebitoId], movCredito)
-				movimiento := asientoContable.CreaMovimiento(el.ValorTotal, dsc, proveedorId, detalleCuenta[cuentasSubgrupo[el.SubgrupoCatalogoId].CuentaDebitoId], movDebito)
-				movimientos = append(movimientos, movimientoR, movimiento)
-
-			} else if el.ValorTotal != originales[idx].ValorTotal {
-
-				tipoMovimiento := movDebito
-				if el.ValorTotal < originales[idx].ValorTotal {
-					tipoMovimiento = movCredito
-				}
-
-				movimiento := asientoContable.CreaMovimiento(math.Abs(el.ValorTotal-originales[idx].ValorTotal), dsc, proveedorId, detalleCuenta[cuentasSubgrupo[el.SubgrupoCatalogoId].CuentaDebitoId], tipoMovimiento)
-				movimientos = append(movimientos, movimiento)
-
-			}
-		}
-
-	}
-
-	for _, el := range actualizarVl {
-		if idx := findElementoInArrayE(originales, el.Id); idx > -1 {
-
-			if detalleCuenta_, err := fillCuentas(detalleCuenta, []string{cuentasSubgrupo[originales[idx].SubgrupoCatalogoId].CuentaCreditoId,
-				cuentasSubgrupo[el.SubgrupoCatalogoId].CuentaCreditoId, cuentasSubgrupo[originales[idx].SubgrupoCatalogoId].CuentaDebitoId, cuentasSubgrupo[el.SubgrupoCatalogoId].CuentaDebitoId}); err != nil {
-			} else {
-				detalleCuenta = detalleCuenta_
-			}
-
-			if el.ValorTotal != originales[idx].ValorTotal {
-
-				tipoMovimientoC := movCredito
-				tipoMovimientoD := movDebito
-
-				if el.ValorTotal < originales[idx].ValorTotal {
-					tipoMovimientoC = movDebito
-					tipoMovimientoD = movCredito
-				}
-
-				movimientoC := asientoContable.CreaMovimiento(math.Abs(el.ValorTotal-originales[idx].ValorTotal), dsc, proveedorId, detalleCuenta[cuentasSubgrupo[el.SubgrupoCatalogoId].CuentaCreditoId], tipoMovimientoC)
-				movimientoD := asientoContable.CreaMovimiento(math.Abs(el.ValorTotal-originales[idx].ValorTotal), dsc, proveedorId, detalleCuenta[cuentasSubgrupo[el.SubgrupoCatalogoId].CuentaDebitoId], tipoMovimientoD)
-				movimientos = append(movimientos, movimientoC, movimientoD)
-			}
-
-		}
-
-	}
-
-	return movimientos, nil
-
-}
-
-// calcularAjusteMediciones Vuelve a generar las novedades y calcula las transacciones contables según las modificaciones que hayan afectado mediciones posteriores aprobadas
-func calcularAjusteMediciones(novedades map[int][]*models.NovedadElemento, sg, vls, mp []*models.DetalleElemento_, org []*models.Elemento) (movimientos []*models.MovimientoTransaccion, novedades_ []*models.NovedadElemento, outputError map[string]interface{}) {
-
-	var (
-		cuentasSubgrupo map[int]*models.CuentaSubgrupo
-		bufferCtas      map[string]*models.CuentaContable
-		movDebito       int
-		movCredito      int
-		terceroUD       int
-	)
-
-	detalleMd := make(map[int]*models.FormatoDepreciacion)
-	novedadesNuevas := make(map[int][]*models.NovedadElemento)
-
-	if db_, cr_, err := parametrosHelper.GetParametrosDebitoCredito(); err != nil {
-		return nil, nil, err
-	} else {
-		movDebito = db_
-		movCredito = cr_
-	}
-
-	if terceroUD_, err := tercerosHelper.GetAllDatosIdentificacion(queryUD + tercerosHelper.GetDocUD()); err != nil {
-		return nil, nil, err
-	} else {
-		terceroUD = terceroUD_[0].TerceroId.Id
-	}
-
-	if cuentasSg, cuentas, err := consultaCuentasMp(novedades, sg, vls, mp, org); err != nil {
-		return nil, nil, err
-	} else {
-		cuentasSubgrupo = cuentasSg
-		bufferCtas = cuentas
-	}
-
-	for key, nv := range novedades {
-
-		var nuevo *models.DetalleElemento_
-		var sgOrg int
-		if idx := findElementoInArrayD(sg, nv[0].ElementoMovimientoId.ElementoActaId); idx > -1 {
-			nuevo = sg[idx]
-			if idx := findElementoInArrayE(org, nv[0].ElementoMovimientoId.ElementoActaId); idx > -1 {
-				sgOrg = org[idx].SubgrupoCatalogoId
-				org = append(org[:idx], org[idx+1:]...)
-			}
-			sg = append(sg[:idx], sg[idx+1:]...)
-		}
-
-		if idx := findElementoInArrayD(vls, nv[0].ElementoMovimientoId.ElementoActaId); idx > -1 {
-			nuevo = vls[idx]
-			vls = append(vls[:idx], vls[idx+1:]...)
-		}
-
-		if idx := findElementoInArrayD(mp, nv[0].ElementoMovimientoId.ElementoActaId); idx > -1 {
-			nuevo = mp[idx]
-			mp = append(mp[:idx], mp[idx+1:]...)
-		}
-
-		for idx, nv_ := range nv {
-
-			if detalleMd[nv_.MovimientoId.Id] == nil {
-				if dt, err := depreciacionHelper.GetDetalleDepreciacion(nv_.MovimientoId.Detalle); err != nil {
-					return nil, nil, err
-				} else {
-					detalleMd[nv_.MovimientoId.Id] = dt
-				}
-			}
-
-			var fCorte time.Time
-			var dpOrg, dpNvo, deltaT float64
-			var novedadNueva *models.NovedadElemento
-			fCorte, _ = time.Parse("2006-01-02", detalleMd[nv_.MovimientoId.Id].FechaCorte)
-			if idx == 0 {
-				dpOrg, _ = depreciacionHelper.CalculaDp(nv_.ElementoMovimientoId.ValorTotal, nv_.ElementoMovimientoId.ValorResidual, nv_.ElementoMovimientoId.VidaUtil, nv_.ElementoMovimientoId.MovimientoId.FechaModificacion, fCorte)
-				dpNvo, deltaT = depreciacionHelper.CalculaDp(nuevo.ValorTotal, nuevo.ValorResidual, nuevo.VidaUtil, nv_.ElementoMovimientoId.MovimientoId.FechaModificacion, fCorte)
-				novedadNueva = generarNovedad(nuevo.ValorTotal-dpNvo, nuevo.ValorResidual, nuevo.VidaUtil-deltaT, nv_)
-			} else {
-				ref, _ := time.Parse("2006-01-02", detalleMd[novedadesNuevas[key][idx-1].MovimientoId.Id].FechaCorte)
-				dpOrg, _ = depreciacionHelper.CalculaDp(nv_.ValorLibros, nv_.ValorResidual, nv_.VidaUtil, ref.AddDate(0, 0, 1), fCorte)
-				dpNvo, deltaT = depreciacionHelper.CalculaDp(novedadesNuevas[key][idx-1].ValorLibros, nuevo.ValorResidual, novedadesNuevas[key][idx-1].VidaUtil, ref.AddDate(0, 0, 1), fCorte)
-				novedadNueva = generarNovedad(novedadesNuevas[key][idx-1].ValorLibros-dpNvo, nuevo.ValorResidual, novedadesNuevas[key][idx-1].VidaUtil-deltaT, nv_)
-			}
-
-			novedadesNuevas[key] = append(novedadesNuevas[key], novedadNueva)
-
-			movimientos = append(movimientos, generaTrContable(dpNvo-dpOrg, detalleMd[nv_.MovimientoId.Id].FechaCorte,
-				nv_.MovimientoId.FormatoTipoMovimientoId.Nombre, movDebito, movCredito, sgOrg, nuevo.SubgrupoCatalogoId, terceroUD, cuentasSubgrupo, bufferCtas)...)
-		}
-	}
-
-	return movimientos, nil, nil
-
-}
-
-// generaTrContable Dado un valor, subgrupo nuevo y original genera la transacción contable.
-func generaTrContable(delta float64, consecutivo, tipoMedicion string, db, cr, sgOriginal, sgNuevo, tercero int, ctasSg map[int]*models.CuentaSubgrupo, ctas map[string]*models.CuentaContable) (movimientos []*models.MovimientoTransaccion) {
-
-	dsc := getDescripcionMovContable(tipoMedicion, consecutivo)
-	if sgOriginal > 0 {
-		if ctasSg[sgOriginal].CuentaCreditoId != ctasSg[sgNuevo].CuentaCreditoId {
-			movimientoR := asientoContable.CreaMovimiento(math.Abs(delta), dsc, tercero, ctas[ctasSg[sgOriginal].CuentaDebitoId], cr)
-			movimiento := asientoContable.CreaMovimiento(math.Abs(delta), dsc, tercero, ctas[ctasSg[sgNuevo].CuentaCreditoId], db)
-			movimientos = append(movimientos, movimientoR, movimiento)
-		} else if delta != 0 {
-			tipoMovimiento := cr
-			if delta < 0 {
-				tipoMovimiento = db
-			}
-
-			movimiento := asientoContable.CreaMovimiento(math.Abs(delta), dsc, tercero, ctas[ctasSg[sgNuevo].CuentaCreditoId], tipoMovimiento)
-			movimientos = append(movimientos, movimiento)
-		}
-
-		if ctasSg[sgOriginal].CuentaDebitoId != ctasSg[sgNuevo].CuentaDebitoId {
-			movimientoR := asientoContable.CreaMovimiento(math.Abs(delta), dsc, tercero, ctas[ctasSg[sgOriginal].CuentaCreditoId], db)
-			movimiento := asientoContable.CreaMovimiento(math.Abs(delta), dsc, tercero, ctas[ctasSg[sgNuevo].CuentaDebitoId], cr)
-			movimientos = append(movimientos, movimientoR, movimiento)
-		} else if delta != 0 {
-			tipoMovimiento := db
-			if delta < 0 {
-				tipoMovimiento = cr
-			}
-
-			movimiento := asientoContable.CreaMovimiento(math.Abs(delta), dsc, tercero, ctas[ctasSg[sgNuevo].CuentaDebitoId], tipoMovimiento)
-			movimientos = append(movimientos, movimiento)
-		}
-
-	} else {
-
-		tipoMovimientoC := cr
-		tipoMovimientoD := db
-
-		if delta < 0 {
-			tipoMovimientoC = db
-			tipoMovimientoD = cr
-		}
-
-		movimientoC := asientoContable.CreaMovimiento(math.Abs(delta), dsc, tercero, ctas[ctasSg[sgNuevo].CuentaCreditoId], tipoMovimientoC)
-		movimientoD := asientoContable.CreaMovimiento(math.Abs(delta), dsc, tercero, ctas[ctasSg[sgNuevo].CuentaDebitoId], tipoMovimientoD)
-		movimientos = append(movimientos, movimientoC, movimientoD)
-
-	}
-
-	return movimientos
-}
-
-// generarNovedad Actualiza los valores afectados de una novedad al hacer un ajuste a un elemento
-func generarNovedad(libros, residual, util float64, novedad *models.NovedadElemento) *models.NovedadElemento {
-
-	novedad.ValorLibros = libros
-	novedad.VidaUtil = util
-	novedad.ValorResidual = residual
-
-	return novedad
-
-}
-
-// consultaCuentasMp Consulta las cuentas asignadas a cada subgrupo y su detalle según el tipo de novedad
-func consultaCuentasMp(novedades map[int][]*models.NovedadElemento, sg, vls, mp []*models.DetalleElemento_, org []*models.Elemento) (ctasSg map[int]*models.CuentaSubgrupo, ctas map[string]*models.CuentaContable, outputError map[string]interface{}) {
-
-	var idsD, idsA []int
-	var idD, idA int
-	var ctasD, ctasA map[int]*models.CuentaSubgrupo
-
-	for _, nv := range novedades {
-		var ids []int
-		if len(sg) > 0 {
-			if idx := findElementoInArrayD(sg, nv[0].ElementoMovimientoId.ElementoActaId); idx > -1 {
-				ids = append(ids, sg[idx].SubgrupoCatalogoId)
-				if idx := findElementoInArrayE(org, nv[0].ElementoMovimientoId.ElementoActaId); idx > -1 {
-					ids = append(ids, org[idx].SubgrupoCatalogoId)
-					org = append(org[:idx], org[idx+1:]...)
-				}
-				sg = append(sg[:idx], sg[idx+1:]...)
-			}
-		} else if len(vls) > 0 {
-			if idx := findElementoInArrayD(vls, nv[0].ElementoMovimientoId.ElementoActaId); idx > -1 {
-				ids = append(ids, vls[idx].SubgrupoCatalogoId)
-				vls = append(vls[:idx], vls[idx+1:]...)
-			}
-		} else if len(mp) > 0 {
-			if idx := findElementoInArrayD(mp, nv[0].ElementoMovimientoId.ElementoActaId); idx > -1 {
-				ids = append(ids, mp[idx].SubgrupoCatalogoId)
-				mp = append(mp[:idx], mp[idx+1:]...)
-			}
-		}
-
-		if nv[0].MovimientoId.FormatoTipoMovimientoId.Nombre == "Depreciación" {
-			if idD == 0 {
-				idD = nv[0].MovimientoId.FormatoTipoMovimientoId.Id
-			}
-			idsD = append(idsD, ids...)
-		} else if nv[0].MovimientoId.FormatoTipoMovimientoId.Nombre == "Amortizacion" {
-			if idA == 0 {
-				idA = nv[0].MovimientoId.FormatoTipoMovimientoId.Id
-			}
-			idsA = append(idsA, ids...)
-		}
-
-	}
-
-	if idD > 0 {
-		if ctas, err := getCuentasByMovimientoSubgrupos(idD, idsD); err != nil {
-			return nil, nil, err
-		} else {
-			ctasD = ctas
-		}
-	}
-
-	if idA > 0 {
-		if ctas, err := getCuentasByMovimientoSubgrupos(idA, idsA); err != nil {
-			return nil, nil, err
-		} else {
-			ctasA = ctas
-		}
-	}
-
-	ctasSg = joinMaps(ctasD, ctasA)
-
-	idsCtas := make([]string, 0)
-	for _, ctas := range ctasSg {
-		idsCtas = append(idsCtas, ctas.CuentaDebitoId, ctas.CuentaCreditoId)
-	}
-
-	ctas = make(map[string]*models.CuentaContable)
-	if detalleCuenta_, err := fillCuentas(ctas, idsCtas); err != nil {
-		return nil, nil, err
-	} else {
-		ctas = detalleCuenta_
-	}
-
-	return ctasSg, ctas, nil
-
-}
-
-// getCuentasByMovimientoSubgrupos Retorna las cuentas de cada subgrupo en una estructura para fácil acceso
-func getCuentasByMovimientoSubgrupos(movimientoId int, subgrupos []int) (cuentasSubgrupo map[int]*models.CuentaSubgrupo, outputError map[string]interface{}) {
-
-	cuentasSubgrupo = make(map[int]*models.CuentaSubgrupo)
-
-	query := "limit=-1&fields=CuentaDebitoId,CuentaCreditoId,SubgrupoId&sortby=Id&order=desc&"
-	query += "query=SubtipoMovimientoId:" + strconv.Itoa(movimientoId) + ",Activo:true,SubgrupoId__Id__in:"
-	query += url.QueryEscape(utilsHelper.ArrayToString(subgrupos, "|"))
-	if cuentas_, err := catalogoElementosHelper.GetAllCuentasSubgrupo(query); err != nil {
-		return nil, err
-	} else {
-		for _, cuenta := range cuentas_ {
-			cuentasSubgrupo[cuenta.SubgrupoId.Id] = cuenta
-		}
-	}
-
-	return cuentasSubgrupo, nil
-
-}
-
-func joinMaps(map1, map2 map[int]*models.CuentaSubgrupo) map[int]*models.CuentaSubgrupo {
-
-	if len(map1) > 0 {
-		for sg, ctas := range map2 {
-			map1[sg] = ctas
-		}
-		return map1
-	} else if len(map2) > 0 {
-		for sg, ctas := range map1 {
-			map2[sg] = ctas
-		}
-		return map2
-	}
-
-	return nil
-
-}
-
-// fillCuentas Consulta el detalle de una serie de cuentas
-func fillCuentas(cuentas map[string]*models.CuentaContable, cuentas_ []string) (cuentasCompletas map[string]*models.CuentaContable, outputError map[string]interface{}) {
-
-	funcion := "fillCuentas"
-	defer errorctrl.ErrorControlFunction(funcion+" - Unhandled Error!", "500")
-
-	for _, id := range cuentas_ {
-		if _, ok := cuentas[id]; !ok {
-			if cta_, err := cuentasContablesHelper.GetCuentaContable(id); err != nil {
-				return nil, err
-			} else {
-				cuentas[id] = cta_
-			}
-		}
-	}
-
-	return cuentas, nil
-
-}
-
-// findElementoInArray Retorna la posicion en que se encuentra el id específicado
-func findElementoInArrayD(elementos []*models.DetalleElemento_, id int) (i int) {
-	for i, el_ := range elementos {
-		if int(el_.Id) == id {
-			return i
-		}
-	}
-	return -1
-}
-
-// findElementoInArray Retorna la posicion en que se encuentra el id específicado
-func findElementoInArrayE(elementos []*models.Elemento, id int) (i int) {
-	for i, el_ := range elementos {
-		if int(el_.Id) == id {
-			return i
-		}
-	}
-	return -1
-}
-
-func getTipoComprobanteAjustes() string {
-	return "N20"
-}
-
-func getDescripcionMovContable(tipoMovimiento, consecutivo string) string {
-	return "Ajuste contable " + tipoMovimiento + " " + consecutivo
 }
