+++ helpers/ajustesHelper/ajustes.utils.helper.go
@@ -1,0 +1,156 @@
+package ajustesHelper
+
+import (
+	"math"
+	"net/url"
+	"strconv"
+
+	"github.com/udistrital/arka_mid/helpers/asientoContable"
+	"github.com/udistrital/arka_mid/helpers/catalogoElementosHelper"
+	"github.com/udistrital/arka_mid/helpers/cuentasContablesHelper"
+	"github.com/udistrital/arka_mid/helpers/utilsHelper"
+	"github.com/udistrital/arka_mid/models"
+	"github.com/udistrital/utils_oas/errorctrl"
+)
+
+// generaTrContable Dado un valor, subgrupo nuevo y original genera la transacción contable.
+func generaTrContable(delta float64, consecutivo, tipoMedicion string, db, cr, sgOriginal, sgNuevo, tercero int, ctasSg map[int]*models.CuentaSubgrupo, ctas map[string]*models.CuentaContable) (movimientos []*models.MovimientoTransaccion) {
+
+	dsc := getDescripcionMovContable(tipoMedicion, consecutivo)
+	if sgOriginal > 0 {
+		if ctasSg[sgOriginal].CuentaCreditoId != ctasSg[sgNuevo].CuentaCreditoId {
+			movimientoR := asientoContable.CreaMovimiento(math.Abs(delta), dsc, tercero, ctas[ctasSg[sgOriginal].CuentaDebitoId], cr)
+			movimiento := asientoContable.CreaMovimiento(math.Abs(delta), dsc, tercero, ctas[ctasSg[sgNuevo].CuentaCreditoId], db)
+			movimientos = append(movimientos, movimientoR, movimiento)
+		} else if delta != 0 {
+			tipoMovimiento := cr
+			if delta < 0 {
+				tipoMovimiento = db
+			}
+
+			movimiento := asientoContable.CreaMovimiento(math.Abs(delta), dsc, tercero, ctas[ctasSg[sgNuevo].CuentaCreditoId], tipoMovimiento)
+			movimientos = append(movimientos, movimiento)
+		}
+
+		if ctasSg[sgOriginal].CuentaDebitoId != ctasSg[sgNuevo].CuentaDebitoId {
+			movimientoR := asientoContable.CreaMovimiento(math.Abs(delta), dsc, tercero, ctas[ctasSg[sgOriginal].CuentaCreditoId], db)
+			movimiento := asientoContable.CreaMovimiento(math.Abs(delta), dsc, tercero, ctas[ctasSg[sgNuevo].CuentaDebitoId], cr)
+			movimientos = append(movimientos, movimientoR, movimiento)
+		} else if delta != 0 {
+			tipoMovimiento := db
+			if delta < 0 {
+				tipoMovimiento = cr
+			}
+
+			movimiento := asientoContable.CreaMovimiento(math.Abs(delta), dsc, tercero, ctas[ctasSg[sgNuevo].CuentaDebitoId], tipoMovimiento)
+			movimientos = append(movimientos, movimiento)
+		}
+
+	} else {
+
+		tipoMovimientoC := cr
+		tipoMovimientoD := db
+
+		if delta < 0 {
+			tipoMovimientoC = db
+			tipoMovimientoD = cr
+		}
+
+		movimientoC := asientoContable.CreaMovimiento(math.Abs(delta), dsc, tercero, ctas[ctasSg[sgNuevo].CuentaCreditoId], tipoMovimientoC)
+		movimientoD := asientoContable.CreaMovimiento(math.Abs(delta), dsc, tercero, ctas[ctasSg[sgNuevo].CuentaDebitoId], tipoMovimientoD)
+		movimientos = append(movimientos, movimientoC, movimientoD)
+
+	}
+
+	return movimientos
+
+}
+
+// getCuentasByMovimientoSubgrupos Retorna las cuentas de cada subgrupo en una estructura para fácil acceso
+func getCuentasByMovimientoSubgrupos(movimientoId int, subgrupos []int) (cuentasSubgrupo map[int]*models.CuentaSubgrupo, outputError map[string]interface{}) {
+
+	funcion := "getCuentasByMovimientoSubgrupos"
+	defer errorctrl.ErrorControlFunction(funcion+" - Unhandled Error!", "500")
+
+	cuentasSubgrupo = make(map[int]*models.CuentaSubgrupo)
+
+	query := "limit=-1&fields=CuentaDebitoId,CuentaCreditoId,SubgrupoId&sortby=Id&order=desc&"
+	query += "query=SubtipoMovimientoId:" + strconv.Itoa(movimientoId) + ",Activo:true,SubgrupoId__Id__in:"
+	query += url.QueryEscape(utilsHelper.ArrayToString(subgrupos, "|"))
+	if cuentas_, err := catalogoElementosHelper.GetAllCuentasSubgrupo(query); err != nil {
+		return nil, err
+	} else {
+		for _, cuenta := range cuentas_ {
+			cuentasSubgrupo[cuenta.SubgrupoId.Id] = cuenta
+		}
+	}
+
+	return cuentasSubgrupo, nil
+
+}
+
+func joinMaps(map1, map2 map[int]*models.CuentaSubgrupo) map[int]*models.CuentaSubgrupo {
+
+	if len(map1) > 0 {
+		for sg, ctas := range map2 {
+			map1[sg] = ctas
+		}
+		return map1
+	} else if len(map2) > 0 {
+		for sg, ctas := range map1 {
+			map2[sg] = ctas
+		}
+		return map2
+	}
+
+	return nil
+
+}
+
+// fillCuentas Consulta el detalle de una serie de cuentas
+func fillCuentas(cuentas map[string]*models.CuentaContable, cuentas_ []string) (cuentasCompletas map[string]*models.CuentaContable, outputError map[string]interface{}) {
+
+	funcion := "fillCuentas"
+	defer errorctrl.ErrorControlFunction(funcion+" - Unhandled Error!", "500")
+
+	for _, id := range cuentas_ {
+		if _, ok := cuentas[id]; !ok {
+			if cta_, err := cuentasContablesHelper.GetCuentaContable(id); err != nil {
+				return nil, err
+			} else {
+				cuentas[id] = cta_
+			}
+		}
+	}
+
+	return cuentas, nil
+
+}
+
+// findElementoInArray Retorna la posicion en que se encuentra el id específicado
+func findElementoInArrayD(elementos []*models.DetalleElemento_, id int) (i int) {
+	for i, el_ := range elementos {
+		if int(el_.Id) == id {
+			return i
+		}
+	}
+	return -1
+}
+
+// findElementoInArray Retorna la posicion en que se encuentra el id específicado
+func findElementoInArrayE(elementos []*models.Elemento, id int) (i int) {
+	for i, el_ := range elementos {
+		if int(el_.Id) == id {
+			return i
+		}
+	}
+	return -1
+}
+
+func getTipoComprobanteAjustes() string {
+	return "N20"
+}
+
+func getDescripcionMovContable(tipoMovimiento, consecutivo string) string {
+	return "Ajuste contable " + tipoMovimiento + " " + consecutivo
+}
